% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrimleyCalculator.R
\name{BrimleyCalculator}
\alias{BrimleyCalculator}
\title{Calculate Brimley/Cocoon line statistics}
\usage{
BrimleyCalculator(birthdate)
}
\value{
date person will cross Brimley/Cocoon line
}
\description{
Calculate Brimley/Cocoon line statistics
}
\examples{
results <- BrimleyCalculator(as.Date("1973-12-18"))
}
