% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrimleyCalculator.R
\name{BrimleyCalculator}
\alias{BrimleyCalculator}
\title{Calculate Brimley/Cacoon line statistics}
\usage{
BrimleyCalculator(bdate)
}
\arguments{
\item{birtdate}{Date of person}
}
\value{
date person will cross Brimley/Cacoon line
}
\description{
Calculate Brimley/Cacoon line statistics
}
\examples{
results <- BrimleyCalculator(as.Date("1973-12-18"))
}
